requirements.txt
flask==2.3.3
gunicorn==21.2.0
flask-wtf==1.1.1
python-dotenv==1.0.0
nltk==3.8.1
pandas==2.1.0
pytest==7.4.0